# Documentation/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ECMA Events
  description: >-
    API da aplicação gerida pela companhia de marketing ECMA para a gestão da
    organização de eventos com a colaboração das suas companhias parceiras.
servers:
  - url: 'http://localhost:8080/'
...
paths:
...
  /partners:
    get:
      tags:
        - Partner Company
      description: Obter todas as empresas parceiras
      operationId: getPartnerCompanies
      parameters:
        - name: criteria
          in: query
          description: Filtrar as empresas que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com a informação sobre as empresas parceiras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartnerCompany'
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Partner Company
      description: Adicionar uma companhia parceira ao sistema
      operationId: addPartnerCompany
      requestBody:
        description: Todos os necessários dados da empresa
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerCompany'
      responses:
        '200':
          description: Empresa adicionada com sucesso
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/partners/{id}':
    get:
      tags:
        - Partner Company
      description: 'Obter a empresa identificada pelo {id}'
      operationId: getPartnerCompany
      parameters:
        - name: id
          in: path
          description: Identificador da empresa a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Informação da empresa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerCompany'
        '404':
          description: Empresa não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Partner Company
      description: 'Atualizar a empresa identificada pelo {id}'
      operationId: updatePartnerCompany
      parameters:
        - name: id
          in: path
          description: Identificador da empresa a remover
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          description: Objeto da empresa que é atualizada
          required: true
          schema:
            $ref: '#/components/schemas/PartnerCompany'
      responses:
        '200':
          description: Empresa atualizada
        '400':
          description: Id especificado inválido
        '404':
          description: Empresa não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Partner Company
      description: 'Remover do sistema a empresa identificada pelo {id}'
      operationId: deletePartnerCompany
      parameters:
        - name: id
          in: path
          description: Identificador da empresa a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Empresa removida
        '404':
          description: Empresa não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/partners/{id}/employees':
    get:
      tags:
        - Partner Company
      description: 'Obter todos os empregados da empresa parceira identificada pelo {id}'
      operationId: getCompanyEmployees
      parameters:
        - name: id
          in: path
          description: Identificador da empresa parceira
          required: true
          schema:
            type: integer
            format: int64
        - name: criteria
          in: query
          description: Filtrar os empregados que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com as informações dos empregados da empresa em questão
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '404':
          description: Empresa não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Partner Company
      description: 'Adicionar um empregado à companhia parceira identificada pelo {id}'
      operationId: addEmployee
      parameters:
        - name: id
          in: path
          description: Identificador da empresa parceira
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Todos os necessários dados do empregado
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Empregado adicionado com sucesso
        '404':
          description: Empresa não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PartnerCompany:
      required:
        - name
        - city
        - zipCode
        - streetAddress
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          example: ECMA
        city:
          type: string
          example: Lisbon
        zipCode:
          type: string
          example: 1000-261
        streetAddress:
          type: string
          example: Long Street
        phone:
          type: integer
          format: int64
          example: 212121212
        emailAddress:
          type: string
          example: info@ecma.com
        fax:
          type: string
          example: 211212121
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
          readOnly: true
    Employee:
      required:
        - userId
        - companyId
        - city
        - zipCode
        - streetAddress
        - cellPhone
        - homePhone
        - gender
        - salary
        - birthday
        - job
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        userId:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        city:
          type: string
          example: Lisbon
        zipCode:
          type: string
          example: 1000-260
        streetAddress:
          type: string
          example: Short Street
        cellPhone:
          type: integer
          format: int64
          example: 919191919
        homePhone:
          type: integer
          format: int64
          example: 211111111
        gender:
          type: string
          enum:
            - M
            - F
        salary:
          type: integer
          format: int64
          example: 5000
          minimum: 0
        birthday:
          type: string
          example: '2018-01-01'
          format: date
        job:
          type: string
          example: CEO
        isAdmin:
          type: boolean
          default: false
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string